Add files via upload
In this Jupiter file, we want to download the tweets and do some initial processing on them. There are two outputs for this code: first, there is a file entitled 'tweet_information.csv'; and second, there is a file entitled 'most_words_list_result.csv'. First file contains the information about tweets (such as the text, ID, date, etc.).
Second file contains the frequency of "the most repetitious words in each class" in all classes. For example, if the frequency of the word 'the' is 27, it means that this word has been repetitious in 27 classes; and if the frequency of the word 'close' is 1, it means that this word has been repetitious in 1 class. Being repetitious means being one of the 266 repetitious words, in one class or more.
Also, other advanced processing on data, will be applied in the next Jupiter file.