Tweets with some initial processing
This file in entitled 'tweet_information.csv' and contains 7 columns about tweets of Twitter users. The number of tweets is 98681, and all of them has been collected in March 2021, by our team. The columns in this file are ID, user name, date of creation, text of tweet, length of text of tweet, the target tag, and the kind of data collecting filtering (all of them are 'filter:retweets'). Some initial processing has been performed on the tweets, such as lowercasing the text, eliminating emojis and replicas, etc. This file, is one of the outputs of "Downloading_and_init_processing.ipynb" file. Other processing will be considered in the next Jupiter file.