Frequency of the most repetitious words
The file entitled 'most_words_list_result.csv' file is one of the outputs of 'Downloading_and_init_processing.ipynb' code. It contains 1785 of the most repetitious words in tweets, and their frequencies among the classes. At first, we collected the tweets for 27 classes; then we looked for 266 of the most repetitious words in each class; then we put all these 27*266 words together (some of these are repeated of course); finally we had 1785 unique words. Some of these words have been repetitious in 2 classes (or more). That's why we have written the frequencies beside each word. For example, the word 'the' have been repetitious in all classes; so the frequency of this word is 27 (it means that 'the' is one of the 266 repetitious words in all classes).