Advanced processing on data
This is the code that will make the final file of data. It has 3 outputs. First, the final data which is entitled 'data_clean.csv'. Second, the file of binary string of each word or the Space (there are 1785 word and the Space); it relates each of these 1786, to its binary string; each binary srting has eleven digits of 0 and 1. Third, 100 rows of 'data_clean.csv' are chosen to make 'data_clean_sample_100.csv' as the sample data.