Final file of data*
*NOTE: The original size of this file is about 85MB and has 98681 rows. But, we put a smaller file here with 25000 rows, because of 25MB limit of GitHub. If you want to see the larger file, please tell us. Also, the description below, is for the larger file: 
This is the final file of data. It has 98681 rows, and 452 columns. The first 451 columns are the features, and the last column is the label of each class. To make this file, we first related each of 1785 common words to a binary string. For example, the binary string of 'desert' is '11011110110'. The maximum length of each tweet is 41 words. So, each tweet has 11*41=451 digits; we concatenated these 451 digits to make a single string for each tweet. Please note that if the length of a tweet was less than 41, we added Space character; the binary string for each Space is '00000000000'. Also note that OneHot method did not work well for these data, because the number of items for each array (for each word) was larger than 1800; thus, we had some problems about the dimension of data and saving the features. That's why we used binary strings instead of OneHot. So, we have had a map from 41 words, to a binary vector with 451 elements. The last column is for the label of each class. We have 27 classes; so the labels are from 0 to 26. These labels, are the target tag of each tweet.