Binary string for each word
This file contains the binary number for each word. There were 1785 most frequent words. So, we encode each of them by a relating a binary number to each. Also note that the binary number of Space character is '00000000000', but here in this file, first 0 digits has not been shown. These binary numbers will be used for digit encoding of each word.